requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
SQLAlchemy==2.0.31
psycopg2-binary
python-dotenv